separate general logic from json resolver
